+++ src/main/java/techcourse/myblog/dto/request/UserDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.dto.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
